requirements.txt
discord;
os;
youtubesearchpython;
pafy;
python-vlc;